Use i64 for amount

+++ modules/meteroid/crates/meteroid-repository/refinery/migrations/V0025__invoice_add_amount.sql
@@ -1,3 +1,2 @@
--- Modify "customer" table
 ALTER TABLE "invoice"
-    ADD COLUMN "amount_cents" INTEGER;
+    ADD COLUMN "amount_cents" BIGINT;
